chore(v0.4.5): mypy 严格化 check_untyped_defs=true (TD-071 RESOLVED)
启用 pyproject.toml mypy check_untyped_defs=true，修复 9 个隐藏类型错误:
- animation_system.py:79 — CONFIGS dict 推断为 dict[str, int] (bool 并入 int) → 显式标注为 dict[str, int | bool] + bool() 转换
- cc2_combat_effects.py:150 — Particle.size: int 但 smoke 扩散 += 0.1 产生 float → 改为 size: float
- environment_renderer.py:97 — 函数声明 -> tuple[Surface, Surface] 但缓存可能为 None → 改为 tuple[Surface | None, Surface | None]
- particle_pool.py:44 — _pool: list[object] 导致属性访问报错 → 改为 list[Any]
- tutorial_system.py:277/283/287/308 — 4 处 _font_*.render() 在 lazy-init 后未窄化 None → 添加 assert 窄化
- interaction_controller_protocol.py — Protocol 缺 clear_selection 方法 → 补充接口

tests.* override 保留 check_untyped_defs=false（测试代码不强制严格类型检查）

验证: mypy 0 errors (389 files), ruff 0, unit 4611 passed/2 skipped, 0 回归

+++ src/pycc2/presentation/rendering/cc2_combat_effects.py
@@ -26,7 +26,7 @@ class EnhancedParticleSystem:
         vy: float
         life: int
         max_life: int
-        size: int
+        size: float
         color: tuple[int, int, int]
         particle_type: str  # "fire", "smoke", "debris", "spark"
